Added command line argument support for file and directory processing
Enhanced the script to support processing of individual files and specific directories through command line arguments. The script can now be run in three different ways:

1. With a file path as an argument, it processes just that file.
2. With a directory path as an argument, it processes all the suitable files in that directory and its subdirectories.
3. Without any argument, it processes the current directory (the script's working directory).

Also, modified the script to always look for the `.embedding_ignore` file in the script's directory, no matter where the script is executed from.

This update provides greater flexibility for users in specifying the scope of file processing for generating embeddings using the OpenAI API.